requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
psycopg2-binary
